working on file manipulation and users

+++ src/main/java/models/Local_File.java
@@ -10,11 +10,14 @@ import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 import java.util.zip.ZipOutputStream;
 
 public class Local_File implements File_Manipulation_Interface {
+
+    private String storage_path = null;
 
     public Local_File() {
 
@@ -23,24 +26,18 @@ public class Local_File implements File_Manipulation_Interface {
     @Override
     public void create_file(String path, String file_name) throws Create_File_Exception, Invalid_Path_Exception {
 
-        if(!path.isEmpty()) {
-
+        if(!path.isEmpty() && path.startsWith(storage_path)) {
             File file = new File(path + File.separator + file_name);
-
             if(!file.exists()) {
-
                 try {
                     file.createNewFile();
+                    System.out.println("File \"" + file.getName() + "\" successfully created.");
                 } catch (Exception e) {
                     throw new Create_File_Exception();
                 }
-
-                System.out.println("File successfully created.");
-
             } else {
                 throw new Create_File_Exception();
             }
-
         } else {
             throw new Invalid_Path_Exception();
         }
@@ -50,37 +47,16 @@ public class Local_File implements File_Manipulation_Interface {
     @Override
     public void delete_file(String path) throws Delete_Exception {
 
-        if(!path.isEmpty()) {
-
+        if(!path.isEmpty() && path.startsWith(storage_path)) {
             File file = new File(path);
             if(file.exists() && file.isFile()) {
                 file.delete();
+                System.out.println("File \"" + file.getName() + "\" successfully deleted.");
             } else {
                 throw new Delete_Exception();
             }
-
-            System.out.println("File successfully deleted.");
-
         } else {
             throw new Delete_Exception();
-        }
-
-    }
-
-    private void copy_file(File source, File dest) throws IOException {
-        InputStream is = null;
-        OutputStream os = null;
-        try {
-            is = new FileInputStream(source);
-            os = new FileOutputStream(dest);
-            byte[] buffer = new byte[1024];
-            int length;
-            while ((length = is.read(buffer)) > 0) {
-                os.write(buffer, 0, length);
-            }
-        } finally {
-            is.close();
-            os.close();
         }
 
     }
@@ -88,25 +64,20 @@ public class Local_File implements File_Manipulation_Interface {
     @Override
     public void download_file(String source, String destination) throws Download_Exception {
 
-        if(!source.isEmpty() && !destination.isEmpty()) {
-
+        if(!source.isEmpty() && !destination.isEmpty() && destination.startsWith(storage_path)) {
             File file = new File(source);
             if(file.exists() && file.isFile()) {
-
                 File downloaded_file = new File(destination + File.separator + file.getName());
                 try {
                     downloaded_file.createNewFile();
                     Files.copy(Paths.get(file.getAbsolutePath()), Paths.get(destination + File.separator + file.getName()), StandardCopyOption.REPLACE_EXISTING);
+                    System.out.println("File " + file.getName() + " successfully copied to \"" + destination + "\".");
                 } catch (Exception e) {
                     throw new Download_Exception();
                 }
-
-                System.out.println("File successfully downloaded to " + destination + ".");
-
             } else {
                 throw new Download_Exception();
             }
-
         } else {
             throw new Download_Exception();
         }
@@ -114,15 +85,15 @@ public class Local_File implements File_Manipulation_Interface {
     }
 
     @Override
-    public void download_multiple_files(ArrayList<File> files, String source, String destination) throws Download_Multiple_Exception {
+    public void download_multiple_files(List<File> files, String source, String destination) throws Download_Multiple_Exception {
 
-        if (!source.isEmpty() && !destination.isEmpty()){
-
+        if (!source.isEmpty() && !destination.isEmpty() && source.startsWith(storage_path) && destination.startsWith(storage_path)){
             for (File file : files){
                 if (file.exists() && file.isFile()){
                     try{
                         try {
                             Files.copy(Paths.get(file.getAbsolutePath()), Paths.get(destination + File.separator + file.getName()), StandardCopyOption.REPLACE_EXISTING);
+                            System.out.println("Files are successfully copied to \"" + destination + "\".");
                         } catch (Exception e) {
                             throw new Download_Multiple_Exception();
                         }
@@ -133,8 +104,6 @@ public class Local_File implements File_Manipulation_Interface {
                     continue;
                 }
             }
-            System.out.println("Files are successfully downloaded to " + destination + ".");
-
         } else {
             throw new Download_Multiple_Exception();
         }
@@ -143,14 +112,17 @@ public class Local_File implements File_Manipulation_Interface {
 
     @Override
     public void upload_file(String source, String destination) throws Upload_Exception {
-
-        //download_file(destination, source);
+        //In local, it's same as download_file.
     }
 
     @Override
-    public void upload_multiple_files(ArrayList<File> files, String source, String destination) throws Upload_Multiple_Exception {
+    public void upload_multiple_files(List<File> list, String s, String s1) throws Upload_Multiple_Exception {
+        //In local, it's same as download_multiple_file.
+    }
 
-
+    @Override
+    public void upload_multiple_archived_files(List<File> list, String s, String s1) throws Upload_Multiple_Archives_Exception {
+        //In local, it's same as download_multiple_archived_files.
     }
 
     @Override
@@ -159,20 +131,27 @@ public class Local_File implements File_Manipulation_Interface {
         if(!path.isEmpty() && !archive_name.isEmpty()) {
             File file = new File(path);
             if(file.isFile() && file.exists()) {
-
                 try {
-                    FileInputStream in = new FileInputStream(file.getAbsolutePath());
-                    ZipOutputStream out = new ZipOutputStream(new FileOutputStream("C:\\Users\\pavlenski\\Desktop\\SOFT KOMP\\malina\\" + archive_name + ".zip"));
-                    out.putNextEntry(new ZipEntry(file.getName()));
+                    if (!archive_name.endsWith(".zip") && !archive_name.endsWith(".rar")) archive_name += ".zip";
 
-                    byte[] b = new byte[1024];
-                    int count;
+                    FileOutputStream fos = new FileOutputStream(path.substring(0, path.length() - file.getName().length()) +  archive_name);
+                    ZipOutputStream zipOut = new ZipOutputStream(fos);
 
-                    while ((count = in.read(b)) > 0) {
-                        out.write(b, 0, count);
+                    FileInputStream fis = new FileInputStream(file);
+                    ZipEntry zipEntry = new ZipEntry(file.getName());
+
+                    zipOut.putNextEntry(zipEntry);
+
+                    byte[] bytes = new byte[1024];
+                    int length;
+
+                    while((length = fis.read(bytes)) >= 0) {
+                        zipOut.write(bytes, 0, length);
                     }
-                    out.close();
-                    in.close();
+
+                    zipOut.close();
+                    fis.close();
+                    fos.close();
                 } catch (Exception e) {
                     throw new Archive_Exception();
                 }
@@ -182,29 +161,21 @@ public class Local_File implements File_Manipulation_Interface {
         } else {
             throw new Archive_Exception();
         }
-    }
-
-    @Override
-    public void upload_multiple_archived_files(ArrayList<File> list, String s, String s1) throws Upload_Multiple_Archives_Exception {
 
     }
 
     @Override
     public void move_file(String source, String destination) throws Move_Exception {
 
-        if(!source.isEmpty() && !destination.isEmpty()) {
-
+        if(!source.isEmpty() && !destination.isEmpty() && destination.startsWith(storage_path)) {
             File file = new File(source);
             if(file.exists() && file.isFile()){
-
                 try {
                     Files.move(Paths.get(file.getAbsolutePath()), Paths.get(destination + File.separator + file.getName()), StandardCopyOption.REPLACE_EXISTING);
+                    System.out.println("File " + file.getName() + " successfully moved to " + destination + ".");
                 } catch (Exception e) {
                     throw new Move_Exception();
                 }
-
-                System.out.println("File " + file.getName() + " successfully moved to " + destination + ".");
-
             } else {
                 throw new Move_Exception();
             }
@@ -217,24 +188,26 @@ public class Local_File implements File_Manipulation_Interface {
     @Override
     public void rename_file(String path, String new_name) throws Rename_Exception, Invalid_Path_Exception {
 
-        if(!path.isEmpty()) {
-
+        if(!path.isEmpty() && path.startsWith(storage_path)) {
             File file = new File(path);
             if(file.exists() && file.isFile()) {
                 String new_path = path.substring(0, path.length() - file.getName().length());
                 file.renameTo(new File(new_path + new_name));
                 System.out.println("File " + file.getName() + " successfully renamed to " + new_name + ".");
-
             } else {
                 throw new Rename_Exception();
             }
-
         } else {
             throw new Invalid_Path_Exception();
         }
 
     }
 
+    public String getStorage_path() {
+        return storage_path;
+    }
 
-
+    public void setStorage_path(String storage_path) {
+        this.storage_path = storage_path;
+    }
 }
